refactor: move queueserver annotation to its own file
Makes it easier to import
