+++ src/sophys/common/plans/queueserver_annotation.py
@@ -1,0 +1,65 @@
+try:
+    from bluesky_queueserver.manager.annotation_decorator import (
+        parameter_annotation_decorator,
+    )
+except ImportError:
+    import copy
+    import functools
+    import inspect
+
+    def parameter_annotation_decorator(annotation):
+        """
+        Copy of the official one, without validation.
+        See https://github.com/bluesky/bluesky-queueserver/blob/main/bluesky_queueserver/manager/annotation_decorator.py
+        """
+
+        def function_wrap(func):
+            if inspect.isgeneratorfunction(func):
+
+                @functools.wraps(func)
+                def wrapper(*args, **kwargs):
+                    return (yield from func(*args, **kwargs))
+
+            else:
+
+                @functools.wraps(func)
+                def wrapper(*args, **kwargs):
+                    return func(*args, **kwargs)
+
+            # Always create the copy (annotation dictionary may be reused)
+            nonlocal annotation
+            annotation = copy.deepcopy(annotation)
+
+            sig = inspect.signature(func)
+            parameters = sig.parameters
+
+            param_unknown = []
+            if "parameters" in annotation:
+                for p in annotation["parameters"]:
+                    if p not in parameters:
+                        param_unknown.append(p)
+            if param_unknown:
+                msg = (
+                    f"Custom annotation parameters {param_unknown} are not "
+                    f"in the signature of function '{func.__name__}'."
+                )
+                raise ValueError(msg)
+
+            setattr(wrapper, "_custom_parameter_annotation_", annotation)
+
+            return wrapper
+
+        return function_wrap
+
+
+DEFAULT_ANNOTATION = {
+    "parameters": {
+        "detectors": {
+            "convert_device_names": True,
+        },
+        "md": {
+            "convert_device_names": False,
+        },
+    },
+}
+"""A minimal queueserver annotation, removing device name conversion from 'md'."""
